Create README.md
In the Interactive Atlas, when you select "Regional information" then "WGII continental", this produces a global map - a drop-down menu of region names would be handy. When you click over Australia or New Zealand, both Australia and New Zealand are highlighted on the map. Therefore, the region is Australasia, consistent with the name of WGII Chapter 11.  If you then select any of the menu options such as "Time series" or "Table Summary",  the plots or data are incorrectly labeled Australia rather than  Australasia. This needs to be fixed.

For all regions, it's weird that the 95th percentile maximum of daily maximum temperatures is lower for 4 C global warming than 3 C global warming - it should be the opposite!  Why does this occur?